verity: check after activation for any corruption
ActivateByVolumeKey does check for corruption via the dm_status task
after activation, but if it finds that it was corrupted, it only logs
that to stderr and returns 0 anyway.

For cases where corruption can be detected immediately, we want to fail
the mounting, so we add a second check of the same task after activating
the device (or after checking the hash on an existing device), and fail
early.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>