add metadir flag to substitute for /run/atomfs
In some cases, e.g. when guestmounting in a new userns and mountns, but
not chrooted, /run/atomfs may not be writable.

In that situation, use the new --metadir flag to all commands to specify
a replacement for the default /run/atomfs.

This overlaps slightly with the ATOMFS_TEST_RUN_DIR environment
variable, which would have the same effect, but should only be used for
tests, as it is not discoverable.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>