Centralize metadata, move mount logic to atomfs pkg, add tests (#23)
* write molecule config to metadata path

The molecule config contains the OCI path and tag, which is important if
we want to track what container image is mounted at a particular path,
and thus what container might need to stop if a verity error is detected
in one of the atoms in the molecule it's using.

This commit writes that config to a JSON file in the metadata path.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* Move metadir to /run, mount logic into pkg

1 - To make it easy to read the config.json for a given mount,
move the metadata path to
/run/atomfs/meta/$current-mountns/$munged-mountpt-name/

using the current mount NS name in the path means we can track mounting
different images on to the same mountpoint in different mount
namespaces.

2 - move mount logic including the metadata dir logic from
cmd/atomfs/mount.go to atomfs/molecule.go so that users of the package
will also get the same metadir / config.json etc behavior that the
`atomfs` tool does.

As part of this move, we no longer mount an RO overlay in one place and
then either remount another overlay or bind mount to the final target
mountpoint. Instead we build one overlay mount for the mountpoint and
either it has an upperdir/workdir or not.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* add flag to allow missing verity, enforce it

pass through to molecule config, and check to be sure we don't
guestmount and ignore verity data without explicitly saying we want to

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* verify: return error when no squash mounts found

In the guestmount case, we use squashfuse and there is no verity mount
source to check. Treat this as a verify error.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* test: add bats tests for mount

Add a bats tests suite for mounting and for failing to mount when the
images are bad.

Uses the ATOMFS_TEST_RUN_DIR env var to avoid polluting your host's
/run/atomfs/meta dir.

copies the guestmount test from the github yaml into bats and expands it
a bit.

I apologize for the bash quoting situation in the heredoc in the
guestmount tests, forgive me

Missing cases:
- testing `atomfs verify` on bad images:
  requires manufacturing a verity image that will mount OK but has a bad
  block that won't get read until later. I have tested verify with
  mounted bad images that I mounted with a purposely broken atomfs, but
  there should be a better way for CI.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* move github test to use bats

the existing test is now covered there, and we build our own test image
so we can avoid the zothub dep and skopeo dep

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* ensure workdir and upperdir are on same fs

This redefines the --persist argument to point to a directory where
atomfs will create or use both the workdir and the upperdir.

So if you run `atomfs mount --persist=foo/` then the persistent writes
will end up at foo/persist/.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

* add metadir flag to substitute for /run/atomfs

In some cases, e.g. when guestmounting in a new userns and mountns, but
not chrooted, /run/atomfs may not be writable.

In that situation, use the new --metadir flag to all commands to specify
a replacement for the default /run/atomfs.

This overlaps slightly with the ATOMFS_TEST_RUN_DIR environment
variable, which would have the same effect, but should only be used for
tests, as it is not discoverable.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>

---------

Signed-off-by: Michael McCracken <mikmccra@cisco.com>