fix: Export the default extraction policy list and the policy environment variable name
For both squash and erofs, export consts so that callers can
know which environment variable to use to set the policies
(PolicyEnvName), and the default policy list (DefPolicies).

Also add a third const (AllPolicies) which is the full list of
valid extraction policies.  As it happens, those are the same
as the default ones right now, but that needn't always be the
case.

(This addresses one part of issue #29, but doesn't close all of it)

Signed-off-by: Serge Hallyn <serge@hallyn.com>