fix: don't mix up runtimedir and metadir
The runtimedir is e.g. /run/atomfs
The metadir is e.g. /run/atomfs/meta/NSNAME/encoded-path

The RuntimeDir() fn would return runtimedir in most cases,
but in one case, it would return the passed-in metapath.

As a result, it could happen that mount, umount, mount of the
same path would fail the final mount, because it saw that
runtimedir existed, when it should only care if metadir exists.

Signed-off-by: Serge Hallyn <serge@hallyn.com>