fix: Include tag matching for MAIN_VERSION used in --version output
The atomfs --version output should match the build version; currently
it either shows just the git hash, or an incorrect older tag plus
hash.  For example v1.2.0 was tagged back in early 2025, but if you
build tip of the tree, the --version shows 1.1.2-XXXX

$ ./bin/atomfs --version
v1.1.2-9-gb584303

Add --tags to the git-describe output to match tags as well.

Fixes: #45

Signed-off-by: Ryan Harper <rharper@woxford.com>